refactor(eval): move eval services and tests into subpackages
Move 10 eval service files into services/eval/ and 12 eval test files
into tests/eval/. Update all imports and patch() paths across 18 files.
